requirements.txt
Flask
pandas
cassandra-driver
openai
astrapy
python-dotenv